requirements.txt
einops==0.8.1
h5py==3.12.1
lpips==0.1.4
mamba_ssm==2.2.4
matplotlib==3.10.0
nibabel==5.3.2
numpy==2.2.2
opencv_python==4.11.0.86
Pillow==11.1.0
scipy==1.15.1
thop==0.1.1.post2209072238
timm==1.0.14
torch==2.3.0+cu121
torchvision==0.18.0+cu121
